Fix CARD fixed-width ABI declarations

+++ src/static/dolphin/card/CARDDelete.c
@@ -60,8 +60,8 @@ s32 CARDFastDeleteAsync(s32 chan, s32 fileNo, CARDCallback callback) {
     return result;
 }
 
-long CARDFastDelete(long chan, long fileNo) {
-    long result = CARDFastDeleteAsync(chan, fileNo, __CARDSyncCallback);
+s32 CARDFastDelete(s32 chan, s32 fileNo) {
+    s32 result = CARDFastDeleteAsync(chan, fileNo, __CARDSyncCallback);
 
     if (result < 0) {
         return result;
