+++ src/static/dolphin/card/CARDFormat.c
@@ -143,8 +143,8 @@ s32 CARDFormatAsync(s32 chan, CARDCallback callback) {
 
 // @hack
 #pragma force_active on
-long CARDFormat(long chan) {
-    long result = __CARDFormatRegionAsync(chan, OSGetFontEncode(), __CARDSyncCallback);
+s32 CARDFormat(s32 chan) {
+    s32 result = __CARDFormatRegionAsync(chan, OSGetFontEncode(), __CARDSyncCallback);
 
     if (result < 0) {
         return result;
